+++ R/count_bindings.R
@@ -585,12 +585,58 @@ set_denoms_by.count_layer <- function(e, ...) {
   assert_that(all(dots_chr %in% c(by_, cols_, treat_var_, target_var_)),
               msg = "A denom_by wasn't found as a grouping variable in the layer/table.")
 
+  if(length(target_var) == 2) {
+    assert_that(!(as_name(target_var[[2]]) %in% dots_chr),
+                msg = "You can not pass the second variable in `vars` as a denominator.")
+  }
+
   # If the row variable is here, rename it to summary_var
   if(as_name(target_var[[1]]) %in% dots_chr) {
-    dots[[which(as_name(target_var[[1]]) %in% dots_chr)]] <- quo(summary_var)
+    dots[dots_chr %in% as_name(target_var[[1]])] <- quos(summary_var)
   }
 
   env_bind(e, denoms_by = dots)
+
+  e
+}
+
+#' Select levels to keep in a count layer
+#'
+#' @description
+#' In certain cases you only want a layer to include certain values of a factor.
+#' The `keep_levels()` function allows you to pass character values to be
+#' included in the layer. The others are ignored.
+#' **NOTE: Denominator calculation is unaffected by this function, see the
+#' examples on how to include this logic in your percentages'**
+#'
+#' @param e A \code{count_layer} object
+#' @param ... Character values to count int he layer
+#'
+#' @return The modified Tplyr layer object
+#' @export
+#'
+#' @examples
+#' library(dplyr)
+#' mtcars <- mtcars %>%
+#'   mutate_all(as.character)
+#'
+#' t <- tplyr_table(mtcars, gear) %>%
+#'   add_layer(
+#'     group_count(cyl) %>%
+#'       keep_levels("4", "8") %>%
+#'       set_denom_where(cyl %in% c("4", "8"))
+#'  ) %>%
+#'  build()
+#'
+keep_levels <- function(e, ...) {
+  dots <- list(...)
+  assert_that(all(map_lgl(dots, is.character)),
+              msg = "must pass character values to `keep_levels`")
+
+  # Create the logic for the filtering. Pull out the second variable if it is a
+  # nested count layer, or the first if it isn't nested
+  env_bind(e, kept_levels = expr(!!tail(env_get(e, "target_var"), 1)[[1]] %in% !!unlist(dots)))
+  env_bind(e, levels_to_keep = dots)
 
   e
 }