+++ tests/testthat/test-count.R
@@ -547,6 +547,17 @@ test_that("nested count layers can accecpt text values in the first variable", {
       group_count(vars(cyl, "Txt"))
     )
   expect_error(build(t2), "Inner layers must be data driven variables")
+
+  mtcars$cyl <- factor(as.character(mtcars$cyl), c("4", "6", "8", "25"))
+  t2 <- tplyr_table(mtcars, gear) %>%
+    add_layer(
+      group_count(vars("all cyl", cyl))
+    ) %>%
+    build()
+
+  expect_equal(t2$var1_3,
+               c("15 (100.0%)", " 1 (  6.7%)", " 2 ( 13.3%)", "12 ( 80.0%)",
+                 " 0 (  0.0%)"))
 })
 
 test_that("Variable names will be coersed into symbols", {
@@ -561,4 +572,162 @@ test_that("Variable names will be coersed into symbols", {
       group_count(vars("all cyl", "cyl"))
     )
   expect_warning(build(t2), "The second target variable has been coerced")
+})
+
+test_that("nested count layers can be build with character value in first position and risk difference", {
+  expect_warning({
+    t1 <- tplyr_table(mtcars, gear) %>%
+      add_layer(
+        group_count(vars("all_cyl", cyl)) %>%
+          add_risk_diff(c("4", "5"))
+      ) %>%
+      build()
+  }, "Chi-squared approximation may be incorrect")
+
+
+  expect_equal(t1$rdiff_4_5, c(" 0.000 ( 0.000,  0.000)",
+                               " 0.267 (-0.380,  0.914)",
+                               " 0.133 (-0.441,  0.707)",
+                               "-0.400 (-0.971,  0.171)"))
+})
+
+test_that("keep_levels works as expeceted", {
+  t1 <- tplyr_table(mtcars, gear) %>%
+    add_layer(
+      group_count(cyl) %>%
+        keep_levels("4", "6") %>%
+        set_format_strings(f_str("xxx (xxx%)", n, pct))
+    ) %>%
+    build()
+  t2 <- tplyr_table(mtcars, gear) %>%
+    add_layer(
+      group_count(vars("all cyl", cyl)) %>%
+        keep_levels("8") %>%
+        set_format_strings(f_str("xxx (xxx%)", n, pct))
+    ) %>%
+    build()
+
+  expect_equal(t1$var1_3, c("  1 (  7%)", "  2 ( 13%)"))
+  expect_equal(dim(t1), c(2, 6))
+  expect_equal(t2$var1_3, c(" 12 ( 80%)", " 12 ( 80%)"))
+  expect_equal(dim(t2), c(2, 8))
+
+  expect_error({
+    t3 <- tplyr_table(mtcars, gear) %>%
+      add_layer(
+        group_count(cyl) %>%
+          keep_levels("10", "20")
+      ) %>%
+      build()
+  }, "Error: level passed to `kept_levels` not found: 10 20")
+
+  mtcars$grp <- paste0("grp.", as.numeric(mtcars$cyl) + rep(c(0, 0.5), 16))
+  t4 <- tplyr_table(mtcars, gear) %>%
+    add_layer(
+      group_count(vars(cyl, grp)) %>%
+        keep_levels("nothere")
+    )
+  expect_error(build(t4), "level passed to `kept_levels` not found: nothere")
+})
+
+test_that("nested count layers can be built with restrictive where logic", {
+  mtcars <- mtcars2
+  mtcars$grp <- paste0("grp.", mtcars$cyl + sample(c(0, 0.5), 32, replace = TRUE))
+
+  t <- tplyr_table(mtcars, gear) %>%
+    add_layer(
+      group_count(vars(cyl, grp), where = grp == "grp.8.5")%>%
+        set_nest_count(TRUE) %>%
+        set_order_count_method('bycount') %>%
+        set_ordering_cols("3")
+    ) %>%
+    build()
+
+  expect_equal(dim(t), c(2, 7))
+
+})
+
+test_that("nested count layers handle `set_denoms_by` as expected", {
+  mtcars <- mtcars2
+  mtcars$grp <- paste0("grp.", mtcars$cyl + rep(c(0, 0.5), 16))
+
+  expect_error({
+    t1 <- tplyr_table(mtcars, gear) %>%
+      add_layer(
+        group_count(vars(cyl,grp)) %>%
+          set_denoms_by(grp)
+      )
+  }, "You can not pass the second variable in")
+
+  t2 <- tplyr_table(mtcars, gear) %>%
+    add_layer(
+      group_count(vars(cyl,grp)) %>%
+        set_denoms_by(cyl)
+    ) %>%
+    build()
+
+  expect_equal(t2$var1_3,
+               c(" 1 (  9.1%)", " 1 (  9.1%)", " 0 (  0.0%)", " 2 ( 28.6%)",
+                 " 0 (  0.0%)", " 2 ( 28.6%)", "12 ( 85.7%)", " 7 ( 50.0%)",
+                 " 5 ( 35.7%)"))
+
+  t3 <- tplyr_table(mtcars, gear) %>%
+    add_layer(
+      group_count(vars(cyl,grp)) %>%
+        set_denoms_by(cyl, gear)
+    ) %>%
+    build()
+
+  expect_equal(t3$var1_3,
+               c(" 1 (100.0%)", " 1 (100.0%)", " 0 (  0.0%)", " 2 (100.0%)",
+                 " 0 (  0.0%)", " 2 (100.0%)", "12 (100.0%)", " 7 ( 58.3%)",
+                 " 5 ( 41.7%)"))
+
+
+
+})
+
+test_that("test IBM rounding option", {
+  row_num <- seq(1:2000)
+  trta = ifelse(row_num <= 1000, "Placebo", "ThisDrg")
+  gender = ifelse(between(row_num, 1, 485), "F",
+                  ifelse(between(row_num, 1001, 1525), "F", "M"))
+  adsl <- tibble(trta, gender)
+
+  tabl <- tplyr_table(adsl, trta) %>%
+    add_total_group(group_name = "Total") %>%
+    add_layer(
+      group_count(gender, by = "Gender")   %>%
+        set_format_strings(f_str("xxx (xxx%)", n, pct))
+    ) %>%
+    build()
+
+  expect_equal(tabl$var1_Placebo, c("485 ( 48%)", "515 ( 52%)"))
+
+  options(tplyr.IBMRounding = TRUE)
+
+  tabl2 <- tplyr_table(adsl, trta) %>%
+    add_total_group(group_name = "Total") %>%
+    add_layer(
+      group_count(gender, by = "Gender")  %>%
+        set_format_strings(f_str("xxx (xxx%)", n, pct))
+    )
+  expect_warning(tabl2 <- build(tabl2), "You have enabled IBM Rounding.")
+
+  expect_equal(tabl2$var1_Placebo, c("485 ( 49%)", "515 ( 52%)"))
+
+  options(tplyr.IBMRounding = FALSE)
+})
+
+test_that("nested count layers will error out if second variable is bigger than the first", {
+  mtcars <- mtcars2
+  mtcars$grp <- paste0("grp.", as.numeric(mtcars$cyl) + rep(c(0, 0.5), 16))
+
+  t <- tplyr_table(mtcars, gear) %>%
+    add_layer(
+      group_count(vars(grp, cyl))
+    )
+
+  expect_error(build(t),
+               "The number of values of your second variable must be greater")
 })