+++ NEWS.md
@@ -1,3 +1,26 @@
+# Tplyr 0.4.0
+
+- Enhancements
+  - In certain cases when creating a count layer, you may only want to keep
+  certain factors from your target dataset. Tplyr now has this functionality
+  built in! With the `keep_levels()` you can define what factors you want to
+  keep in your count layers without having to recode/drop factors at the table
+  level.
+  - Tplyr would normally use the R native rounding method and that is the method
+  we recommend. However, in certain cases you may be trying to match your Tplyr
+  output with SAS. You can set the 'tplyr.IBMRounding' option to TRUE, and Tplyr
+  will simulate IBM rounding.
+  - `set_denoms_by()` has been enhanced for nested count layers. You can now your
+  nested count target variables as denominators.
+  
+- Bug fixes
+  - `add_risk_difference()` would error out when you used it in a nested count
+  layer that had a character value as the first variable.
+  - Nested count layers could not be sorted `bycount` if the layer level where
+  logic caused a value to be droped. This was fixed and tested for future
+  development.
+  
+
 # Tplyr 0.3.1
 
 - Enhancements